making strokeMatches into a function rather than a class
